update and fix

+++ app/src/main/java/com/chtmed/roadrunner/domain/PiCalculator.kt
@@ -2,7 +2,7 @@ package com.chtmed.roadrunner.domain
 
 class PiCalculator {
 
-    fun calculate(iterations: Int = 10_000_000): Double {
+    fun calculate(iterations: Long = 10_000_000): Double {
         var pi = 0.0
         var sign = 1.0
         for (i in 0 until iterations) {
